+++ cinelerra/fileexr.h
@@ -45,7 +45,7 @@ public:
 	int colormodel_supported(int colormodel);
 	int read_frame_header(const char *path);
 	int read_frame(VFrame *frame, VFrame *data);
-	int64_t get_memory_usage();
+	size_t get_memory_usage();
 	int write_frame(VFrame *frame, VFrame *data, FrameWriterUnit *unit);
 	FrameWriterUnit* new_writer_unit(FrameWriter *writer);
